Refact: 설정탭 하나로 합침
