+++ KreamWaffleApp/ViewControllers/My Tab/Subviews/MyTabSharedUIStackVIew.swift
@@ -39,6 +39,10 @@ class MyTabSharedUIStackVIew: UIStackView {
         setUpTitleLabel()
     }
     
+    func changeWishNumberCount(newString: String){
+        self.titleLabel3.text = newString
+    }
+    
     required init(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
